@@ -48,13 +48,13 @@ struct VstSyncData
 	bool isCycle;
 	float cycleStart;
 	float cycleEnd;
-	bool m_playbackJumped;
-	int m_bufferSize;
-	int m_sampleRate;
-	int m_bpm;
+	bool playbackJumped;
+	int bufferSize;
+	int sampleRate;
+	int bpm;
 
 #ifdef VST_SNC_LATENCY
-	float m_latency;
+	float latency;
 #endif
 } ;
 
