@@ -69,7 +69,7 @@ ComboBox::ComboBox( QWidget * _parent, const QString & _name ) :
 
 void ComboBox::selectNext()
 {
-	model()->setInitValue( model()->value() + 1 );
+	model()->setValue(model()->value() + 1);
 }
 
 
@@ -77,7 +77,7 @@ void ComboBox::selectNext()
 
 void ComboBox::selectPrevious()
 {
-	model()->setInitValue( model()->value() - 1 );
+	model()->setValue(model()->value() - 1);
 }
 
 
@@ -221,7 +221,7 @@ void ComboBox::wheelEvent( QWheelEvent* event )
 	if( model() )
 	{
 		const int direction = (event->angleDelta().y() < 0 ? 1 : -1) * (event->inverted() ? -1 : 1);
-		model()->setInitValue(model()->value() + direction);
+		model()->setValue(model()->value() + direction);
 		update();
 		event->accept();
 	}
@@ -234,7 +234,7 @@ void ComboBox::setItem( QAction* item )
 {
 	if( model() )
 	{
-		model()->setInitValue( item->data().toInt() );
+		model()->setValue(item->data().toInt());
 	}
 }
 
