@@ -288,6 +288,7 @@ void MidiClip::clearNotes()
 	instrumentTrack()->unlock();
 
 	checkType();
+	updateLength();
 	emit dataChanged();
 }
 
@@ -441,40 +442,47 @@ void MidiClip::checkType()
 }
 
 
-
-
-void MidiClip::saveSettings( QDomDocument & _doc, QDomElement & _this )
+void MidiClip::exportToXML(QDomDocument& doc, QDomElement& midiClipElement, bool onlySelectedNotes)
 {
-	_this.setAttribute( "type", static_cast<int>(m_clipType) );
-	_this.setAttribute( "name", name() );
-	_this.setAttribute("autoresize", QString::number(getAutoResize()));
-	_this.setAttribute("off", startTimeOffset());
+	midiClipElement.setAttribute("type", static_cast<int>(m_clipType));
+	midiClipElement.setAttribute("name", name());
+	midiClipElement.setAttribute("autoresize", QString::number(getAutoResize()));
+	midiClipElement.setAttribute("off", startTimeOffset());
 	
 	if (const auto& c = color())
 	{
-		_this.setAttribute("color", c->name());
+		midiClipElement.setAttribute("color", c->name());
 	}
 	// as the target of copied/dragged MIDI clip is always an existing
 	// MIDI clip, we must not store actual position, instead we store -1
 	// which tells loadSettings() not to mess around with position
-	if( _this.parentNode().nodeName() == "clipboard" ||
-			_this.parentNode().nodeName() == "dnddata" )
+	if (midiClipElement.parentNode().nodeName() == "clipboard" ||
+			midiClipElement.parentNode().nodeName() == "dnddata")
 	{
-		_this.setAttribute( "pos", -1 );
+		midiClipElement.setAttribute("pos", -1);
 	}
 	else
 	{
-		_this.setAttribute( "pos", startPosition() );
+		midiClipElement.setAttribute("pos", startPosition());
 	}
-	_this.setAttribute( "muted", isMuted() );
-	_this.setAttribute( "steps", m_steps );
-	_this.setAttribute( "len", length() );
+	midiClipElement.setAttribute("muted", isMuted());
+	midiClipElement.setAttribute("steps", m_steps);
+	midiClipElement.setAttribute("len", length());
 
 	// now save settings of all notes
 	for (auto& note : m_notes)
 	{
-		note->saveState(_doc, _this);
+		if (!onlySelectedNotes || note->selected())
+		{
+			note->saveState(doc, midiClipElement);
+		}
 	}
+}
+
+
+void MidiClip::saveSettings( QDomDocument & _doc, QDomElement & _this )
+{
+	exportToXML(_doc, _this);
 }
 
 
